List Fragment Working

+++ app/src/main/java/dev/com/sfilizzola/wunderchallenge/models/Car.kt
@@ -1,7 +1,7 @@
 package dev.com.sfilizzola.wunderchallenge.models
 
 data class Car(var address:String,
-               var coordinates: ArrayList<Int>,
+               var coordinates: ArrayList<Double>,
                var engineType:String,
                var exterior:String,
                var fuel:Int,
